third_party/blink/web_tests/svg/custom/document-all-includes-svg-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.all['test1'].length is 2
PASS document.all['test2'].length is 2
PASS document.all['test3'].nodeName is "A"
PASS document.all['test4'].length is 2
PASS document.all['test5'].length is 2
 
 
